% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPgen.R
\name{GPgen}
\alias{GPgen}
\title{Conversion from a long dataframe to a sptm object}
\usage{
GPgen(xd, param, nb, kernel)
}
\arguments{
\item{xd}{Location of the GP prediction}

\item{param}{list of parameters, with the right naming convention}

\item{nb}{Number of GP to be generated}

\item{kernel}{list of kernel used to generate the GPs.}
}
\description{
This function allows you to calculate a Gaussian process kernel
}
\examples{
xd = expand.grid(seq(0,10,length.out = 10),seq(0,10,length.out = 10))
param = list(list(s1=2, s2=5))
nb = 1
kernel = list(k_spatial)
attr(kernel, "type") <- "spatial"
tt = GPgen(xd, param, nb, kernel)
image(matrix(tt$f, ncol = 10,byrow=TRUE))
}
\keyword{Gaussian}
\keyword{process}
