% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPpred.R
\name{GPpred}
\alias{GPpred}
\title{Conversion from a long dataframe to a sptm object}
\usage{
GPpred(xd, x, y, param, kernel)
}
\arguments{
\item{xd}{Location of the GP prediction}

\item{x}{support points}

\item{y}{support values}

\item{param}{list of parameters, with the right naming convention}

\item{kernel}{list of kernel used to generate the GPs.}
}
\description{
This function allows you to calculate a Gaussian process kernel
}
\examples{
x = list(matrix(runif(5,0,10), ncol = 1))
y = runif(5,-2,2)
xd = list(matrix(seq(min(x[[1]])-1,max(x[[1]])+1,length.out = 100), ncol = 1))
param = list(list(q1=1, q2=2))
kernel = list(k_longterm)
t = GPpred(xd, x,y, param, kernel)

xd = list(expand.grid(seq(0,10,length.out = 10),seq(0,10,length.out = 10)))
param = list(list(s1=2, s2=2))
x = list(matrix(runif(20,0,10),ncol = 2))
y = cos(rowSums(x[[1]]))
kernel = list(k_spatial_iso)
attr(kernel, "type") <- "spatial"
tt = GPpred(xd, x,y, param, kernel)
image(matrix(tt$mp, ncol = 10,byrow=TRUE))
}
\keyword{Gaussian}
\keyword{process}
