% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{k_feature}
\alias{k_feature}
\title{Features isotropic kernel}
\usage{
k_feature(f, fp, param = list(f1 = 1, f2 = 1))
}
\arguments{
\item{f}{Features for which the kernel is calculated}

\item{fp}{Features for which the kernel is calculated}

\item{param}{list of parameters specific to the kernel}
}
\description{
Features isotropic kernel
}
\examples{
f = matrix(rnorm(100,0,5), ncol = 5)
fd1 = f[rep(1:nrow(f), nrow(f)),]
fd2 = f[rep(1:nrow(f), each = nrow(f)),]
fdd = cbind(fd1,fd2)
param = list(s1 = 1, s2 = 10)
kern = apply(fdd,1, function(x) k_feature(x[1:ncol(f)],x[(ncol(f)+1):ncol(fdd)],param = param))
k = matrix(kern, ncol = nrow(f))
}
\keyword{Gaussian}
\keyword{kernel}
\keyword{process}

