% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{k_longterm}
\alias{k_longterm}
\title{Long term tempporal kernel}
\usage{
k_longterm(t, tp, param = list(q1 = 1, q2 = 1))
}
\arguments{
\item{t}{Values for which the kernel is calculated}

\item{tp}{Values for which the kernel is calculated}

\item{param}{list of parameters specific to the kernel}
}
\description{
Long term tempporal kernel
}
\examples{
t = seq(0,1,length.out = 10)
param = list(q1=1, q2=1)
k = outer(t,t, k_longterm, param = param)
}
\keyword{Gaussian}
\keyword{kernel}
\keyword{process}

