% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{k_seasonal}
\alias{k_seasonal}
\title{Seasonal tempporal kernel}
\usage{
k_seasonal(t, tp, param = list(q3 = 1, q4 = 1, qs = 1, f = 1))
}
\arguments{
\item{t}{Values for which the kernel is calculated}

\item{tp}{Values for which the kernel is calculated}

\item{param}{list of parameters specific to the kernel}
}
\description{
Seasonal tempporal kernel
}
\examples{
t = seq(0,1,length.out = 10)
param = list(q3=1, q4=1, qs = 1, f = 1)
k = outer(t,t, k_seasonal, param = param)
}
\keyword{Gaussian}
\keyword{kernel}
\keyword{process}

