% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{k_shortterm}
\alias{k_shortterm}
\title{Short term tempporal kernel}
\usage{
k_shortterm(t, tp, param = list(t6 = 1, t7 = 1, t8 = 1))
}
\arguments{
\item{t}{Values for which the kernel is calculated}

\item{tp}{Values for which the kernel is calculated}

\item{param}{list of parameters specific to the kernel}
}
\description{
Short term tempporal kernel
}
\examples{
t = seq(0,1,length.out = 10)
param = list(t6=1, t7=1 ,t8 = 1)
k = outer(t,t, k_shortterm, param = param)
}
\keyword{Gaussian}
\keyword{kernel}
\keyword{process}
