% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{k_spatial_iso}
\alias{k_spatial_iso}
\title{Spatial isotropic kernel}
\usage{
k_spatial_iso(s, sp, param = list(s1 = 1, s2 = 1))
}
\arguments{
\item{s}{Locations for which the kernel is calculated}

\item{sp}{Locations for which the kernel is calculated}

\item{param}{list of parameters specific to the kernel}
}
\description{
Spatial isotropic kernel
}
\examples{
x = seq(0,10,length.out = 4)
xd = expand.grid(x,x)
xd1 = xd[rep(1:nrow(xd), nrow(xd)),]
xd2 = xd[rep(1:nrow(xd), each = nrow(xd)),]
xdd = cbind(xd1,xd2)
param = list(s1 = 1, s2 = 10)
k = matrix(apply(xdd,1, function(x) k_spatial_iso(x[1:2],x[3:4],param = param)), ncol = nrow(xd))
}
\keyword{Gaussian}
\keyword{kernel}
\keyword{process}
