% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{K_wrap}
\alias{K_wrap}
\title{Wrapper for kernel calculation}
\usage{
K_wrap(paramList, y, kernelList, xList)
}
\arguments{
\item{paramList}{list of parameters}

\item{y}{observations}

\item{kernelList}{list of kernels}

\item{xList}{list of features (time, location, features)}
}
\description{
Wrapper for kernel calculation
}
\examples{
kernelList = list(k_spatial_iso)
paramList =  list(list(s1 = 1, s2 = 2), sigma_noise = 0.2)
s = expand.grid(seq(0,10,1), seq(0,10,1))
y = exp( -1/5 * ((s[,1]-5)^2+(s[,2]-5)^2))
xList = list(s)
K = K_wrap(paramList, y, kernelList,xList)
persp(matrix(K[,3], ncol=11))
}
\keyword{Gaussian}
\keyword{likelihood}
\keyword{process}
