% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{optifix}
\alias{optifix}
\title{Optimise with fixed parameters}
\usage{
optifix(
  par,
  fixed,
  fn,
  gr = NULL,
  ...,
  method = c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN"),
  lower = -Inf,
  upper = Inf,
  control = list(),
  hessian = FALSE
)
}
\arguments{
\item{par}{paramters}

\item{fixed}{boolean}

\item{fn}{function}

\item{gr}{gradient of the function}

\item{...}{additional parameters related to the fn and gr functions.}

\item{method}{optimization method}

\item{lower}{lower bounds.}

\item{upper}{upper bounds.}

\item{control}{list of controls.}

\item{hessian}{boolean. Return the Hessian.}
}
\description{
Optimise with fixed parameters
}
\examples{
fn = function(t){(2 - t[1] + t[2])^2 + 2*(3 - t[3])^4}
th = c(1,1,1)
opt = optifix(th,fixed = c(TRUE, FALSE, FALSE), fn = fn, method = "Nelder-Mead")
}
\keyword{fixed}
\keyword{optimization}
\keyword{values.}
\keyword{with}
