% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parToList}
\alias{parToList}
\title{Converts theta to hyperparameters}
\usage{
parToList(theta, kernelList)
}
\arguments{
\item{theta}{parameters to be estimated}

\item{kernelList}{list of kernels}
}
\description{
Converts theta to hyperparameters
}
\examples{
kernelList = list(k_longterm, k_seasonal, k_spatial)
attr(kernelList, "name") <- c("longterm", "seasonal", "spatial")
attr(kernelList, "parameters") <- list(c("q1","q2"), c("q3","q4","qs", "f"), c("s1", "s2"))
theta = c(log(1),log(1),log(1),log(1),log(1),log(1),log(1), log(1), log(0.2))
res = parToList(theta, kernelList)
}
\keyword{Gaussian}
\keyword{process}
\keyword{wrapper}

