% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrap_ll}
\alias{wrap_ll}
\title{Wrapper for GP likelihood maximisation}
\usage{
wrap_ll(theta, y, kernelList, xList)
}
\arguments{
\item{theta}{parameters to be estimated}

\item{y}{observations}

\item{kernelList}{list of kernels}

\item{xList}{list of features}
}
\description{
Wrapper for GP likelihood maximisation
}
\examples{
kernelList = list(k_longterm, k_seasonal, k_spatial)
attr(kernelList, "name") <- c("long term", "seasonal", "spatial")
attr(kernelList, "parameters") <- list(c("q1","q2"), c("q3","q4","qs", "f"), c("s1", "s2"))
theta = c(log(2),log(1),log(1),log(1),log(1),log(1),log(1), log(1), log(0.2))
t  = matrix(rep(seq(0,10,1),11), ncol = 1)
s = expand.grid(seq(0,10,length.out=11), seq(0,10,length.out = 11))
y = cos(rowSums(s) * t)
xList = list(t,t,s)
ll = wrap_ll(theta, y, kernelList,xList)
}
\keyword{Gaussian}
\keyword{likelihood}
\keyword{process}
\keyword{wrapper}

