% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPspt.R
\docType{package}
\name{GPspt-package}
\alias{GPspt}
\alias{GPspt-package}
\title{Generate R documentation from inline comments.}
\description{
GPspt allows you to model spatio-temporal data with Gaussian process models
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/ick003/GPspt}
  \item Report bugs at \url{http://github.com/ick003/GPspt/issues}
}

}
\author{
\strong{Maintainer}: Adrien Ickowicz \email{adrien.ickowicz@csiro.au}

}
